requirements.txt
alembic==1.4.2
app==0.0.1
astroid==2.2.5
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.3
awscli==1.18.74
botocore==1.16.24
Click==7.0
colorama==0.4.3
docutils==0.15.2
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
future==0.17.1
gunicorn==20.0.4
importlib-metadata==1.6.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
lazy-object-proxy==1.4.0
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.8.1
pyasn1==0.4.8
pycodestyle==2.6.0
pycryptodome==3.3.1
PyJWT==1.7.1
pylint==2.3.1
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-editor==1.0.4
python-jose-cryptodome==1.3.2
PyYAML==5.3.1
rsa==3.4.2
s3transfer==0.3.3
six==1.12.0
SQLAlchemy==1.3.17
toml==0.10.1
typed-ast==1.3.5
urllib3==1.25.9
wcwidth==0.2.4
Werkzeug==0.15.2
wrapt==1.11.1
zipp==3.1.0
